+++ Sources/CodexBar/SettingsStore+Defaults.swift
@@ -77,9 +77,6 @@ extension SettingsStore {
         set {
             self.defaultsState.debugDisableKeychainAccess = newValue
             self.userDefaults.set(newValue, forKey: "debugDisableKeychainAccess")
-            if Self.shouldBridgeSharedDefaults(for: self.userDefaults) {
-                Self.sharedDefaults?.set(newValue, forKey: "debugDisableKeychainAccess")
-            }
             KeychainAccessGate.isDisabled = newValue
             self.noteBackgroundWorkSettingsChanged()
         }